add msg_processor agent to master
